requirements.txt
adal==1.2.7
altair==5.5.0
argcomplete==3.5.2
asttokens==3.0.0
attrs==24.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azureml-core==1.59.0
azureml-interpret==1.59.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
dice-ml==0.11
docker==7.1.0
econml==0.15.1
erroranalysis==0.5.4
exceptiongroup==1.2.2
executing==2.1.0
fairlearn==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
fonttools==4.55.3
frozendict==2.4.6
gevent==24.11.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
holidays==0.63
html5lib==1.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
interpret-core==0.5.0
interpret_community==0.31.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
knack==0.12.0
lightgbm==4.5.0
llvmlite==0.39.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_wrappers==0.5.6
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multitasking==0.0.11
narwhals==1.19.1
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==2.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
packaging==24.2
pandas==1.5.3
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
peewee==3.17.8
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
prompt_toolkit==3.0.48
prophet==1.1.6
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyngrok==7.2.2
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rai_core_flask==0.7.6
raiutils==0.4.2
raiwidgets==0.36.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
responsibleai==0.36.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.5.1
scipy==1.13.1
SecretStorage==3.3.3
semver==2.13.0
shap==0.43.0
simple-websocket==1.1.0
six==1.17.0
sklearn-pandas==2.2.0
slicer==0.0.7
smmap==5.0.1
soupsieve==2.6
sparse==0.15.4
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.13.5
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.51
zope.event==5.0
zope.interface==7.2
